Add conversion for badger2 datastore
This change adds functionality to convert an existing datastore to/from a badger2 data store.
