+++ go.mod
@@ -3,16 +3,34 @@ module github.com/ipfs/ipfs-ds-convert
 go 1.12
 
 require (
-	github.com/ipfs/go-datastore v0.4.4
+	github.com/DataDog/zstd v1.4.5 // indirect
+	github.com/alangpierce/go-forceexport v0.0.0-20160317203124-8f1d6941cd75 // indirect
+	github.com/dgraph-io/badger/v2 v2.2007.2 // indirect
+	github.com/dgraph-io/ristretto v0.0.3 // indirect
+	github.com/dgryski/go-farm v0.0.0-20200201041132-a6ae2369ad13 // indirect
+	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/golang/snappy v0.0.2 // indirect
+	github.com/google/uuid v1.1.2 // indirect
+	github.com/ipfs/go-datastore v0.4.5
 	github.com/ipfs/go-ds-badger v0.2.4
-	github.com/ipfs/go-ds-flatfs v0.4.4
+	github.com/ipfs/go-ds-badger2 v0.1.0
+	github.com/ipfs/go-ds-flatfs v0.4.5
 	github.com/ipfs/go-ds-leveldb v0.4.2
 	github.com/ipfs/go-ds-measure v0.1.0
-	github.com/ipfs/go-fs-lock v0.0.4
-	github.com/ipfs/go-ipfs v0.5.0-rc2
-	github.com/ipfs/go-ipfs-config v0.5.2
+	github.com/ipfs/go-fs-lock v0.0.6
+	github.com/ipfs/go-ipfs v0.7.0
+	github.com/ipfs/go-ipfs-config v0.9.0
+	github.com/ipfs/go-log v1.0.4 // indirect
+	github.com/ipfs/go-log/v2 v2.1.1 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
+	github.com/opentracing/opentracing-go v1.2.0 // indirect
 	github.com/pkg/errors v0.9.1
 	github.com/syndtr/goleveldb v1.0.0
 	github.com/urfave/cli v1.22.4
+	github.com/wangjia184/sortedset v0.0.0-20160527075905-f5d03557ba30 // indirect
+	go.uber.org/multierr v1.6.0 // indirect
+	go.uber.org/zap v1.16.0 // indirect
+	golang.org/x/net v0.0.0-20200923182212-328152dc79b1 // indirect
+	golang.org/x/sys v0.0.0-20200923182605-d9f96fdee20d // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
 )
